requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15.1
chardet==5.0.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
packaging==21.3
pip-review==1.2.0
pyparsing==3.0.9
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
